Text curves add

+++ src/twoDimEngine/FillShader.java
@@ -19,8 +19,8 @@ public class FillShader extends Wireframe {
 		int[] x = new int[length];
 		int[] y = new int[length];
 		for (int i = 0; i < length; i++) {
-			x[i] = engine.IntegerCoordX(vec[i].getX());
-			y[i] = engine.IntegerCoordY(vec[i].getY());
+			x[i] = (int) engine.IntegerCoordX(vec[i].getX());
+			y[i] = (int) engine.IntegerCoordY(vec[i].getY());
 		}
 		g.setColor(element2d.getColor(0));
 		g.fillPolygon(x, y, length);
@@ -40,8 +40,8 @@ public class FillShader extends Wireframe {
 	@Override
 	public void paintPoint(Point2D element) {
 		Graphics g = engine.getImageGraphics();
-		int i = engine.IntegerCoordX(element.getVertex(0).getX());
-		int j = engine.IntegerCoordY(element.getVertex(0).getY());
+		int i = (int) engine.IntegerCoordX(element.getVertex(0).getX());
+		int j = (int) engine.IntegerCoordY(element.getVertex(0).getY());
 		g.setColor(element.getColor(0));
 		
 		double radius = element.getRadius();
@@ -49,8 +49,8 @@ public class FillShader extends Wireframe {
 		if(radius == 0.0)
 			g.drawLine(i, j, i, j);
 		else {
-			int rx = engine.IntegerCoordX(element.getRadius()) - engine.IntegerCoordX(0.0);
-			int ry = Math.abs(engine.IntegerCoordY(element.getRadius()) - engine.IntegerCoordY(0.0));
+			int rx = (int) (engine.IntegerCoordX(element.getRadius()) - engine.IntegerCoordX(0.0));
+			int ry = (int) Math.abs(engine.IntegerCoordY(element.getRadius()) - engine.IntegerCoordY(0.0));
 			g.fillOval(i - rx, j - ry, 2 * rx, 2 * ry);
 		}
 	}
